refactor: extract desktop http logic into dedicated service
