+++ src/main/index.ts
@@ -5,14 +5,13 @@ import { BACKEND_URL } from "@shared/const";
 import AutoLaunch from "auto-launch";
 import Database from "better-sqlite3";
 import { drizzle } from "drizzle-orm/better-sqlite3";
-import { app, crashReporter, protocol, session } from "electron";
+import { app, crashReporter, protocol } from "electron";
 import { installExtension, REACT_DEVELOPER_TOOLS } from "electron-devtools-installer";
-import ky from "ky";
-import { Agent, Pool, ProxyAgent } from "undici";
-import { IS_ELECTRON, appVersion } from "./const";
+import { IS_ELECTRON } from "./const";
 import { startInit } from "./init";
 import { DB_FILE_NAME } from "./internal/const";
 import * as schema from "./internal/db/schema";
+import { DesktopHttpService } from "./internal/http";
 import Logger from "./internal/logger";
 import Updater from "./internal/updater";
 import { IPC } from "./ipc";
@@ -55,8 +54,7 @@ const db = drizzle(sqlite, { schema });
 export class NahidaDesktop {
     public initialized: boolean = false;
     public userAgent: string;
-    private cachedAgent: Agent | ProxyAgent | null = null;
-    private cachedProxyHash: string = "";
+    public readonly httpService: DesktopHttpService;
 
     public setting: Setting;
     public readonly ipc: IPC;
@@ -99,6 +97,7 @@ export class NahidaDesktop {
         this.ipc = new IPC(this);
         this.updater = new Updater(this);
         this.logger = new Logger(false, false);
+        this.httpService = new DesktopHttpService(this);
         this.window = {
             main: new MainWindow(this),
             auth: new LoginWindow(this),
@@ -162,132 +161,7 @@ export class NahidaDesktop {
             }
         }
 
-        await this.updateProxy();
-    }
-
-    public async updateProxy() {
-        const proxy = await this.setting.net.getProxy();
-        if (proxy && proxy.type !== "disabled" && proxy.host && proxy.port) {
-            const protocol = proxy.type === "socks5" ? "socks5" : "http";
-            const proxyRules = `${protocol}://${proxy.host}:${proxy.port}`;
-            await app.whenReady();
-            await session.defaultSession.setProxy({ proxyRules });
-            this.logger.info(`Proxy updated: ${proxyRules}`, "NahidaDesktop:updateProxy");
-        } else {
-            await app.whenReady();
-            await session.defaultSession.setProxy({ proxyRules: "" });
-            this.logger.info("Proxy disabled", "NahidaDesktop:updateProxy");
-        }
-    }
-
-    public async getAgent() {
-        const proxy = await this.setting.net.getProxy();
-        const proxyHash = JSON.stringify(proxy);
-
-        if (this.cachedAgent && this.cachedProxyHash === proxyHash) {
-            return this.cachedAgent;
-        }
-
-        if (proxy && proxy.type !== "disabled" && proxy.host && proxy.port) {
-            let url: string;
-            let token: string | undefined;
-
-            if (proxy.type === "socks5") {
-                const auth =
-                    proxy.requiresAuth && proxy.username && proxy.password
-                        ? `${encodeURIComponent(proxy.username)}:${encodeURIComponent(proxy.password)}@`
-                        : "";
-                url = `socks5://${auth}${proxy.host}:${proxy.port}`;
-            } else {
-                url = `http://${proxy.host}:${proxy.port}`;
-                if (proxy.requiresAuth && proxy.username && proxy.password) {
-                    token = `Basic ${Buffer.from(`${proxy.username}:${proxy.password}`).toString("base64")}`;
-                }
-            }
-
-            this.cachedAgent = new ProxyAgent({
-                uri: url,
-                token,
-                factory(origin, options) {
-                    return new Pool(origin, {
-                        ...options,
-                        allowH2: true,
-                    });
-                },
-            });
-        } else {
-            this.cachedAgent = new Agent({
-                factory(origin, options) {
-                    return new Pool(origin, {
-                        ...options,
-                        allowH2: true,
-                    });
-                },
-            });
-        }
-
-        this.cachedProxyHash = proxyHash;
-        return this.cachedAgent;
-    }
-
-    public async getHeaders(url: string) {
-        const token = await this.service.auth.getToken();
-        const prefixs = ["http://localhost", "https://api.nahida.live"];
-        const isNHD = prefixs.some((prefix) => url.startsWith(prefix));
-        return {
-            ...(token && isNHD && { Authorization: `Bearer ${token}` }),
-            Origin: "https://nahida.live",
-            "User-Agent": `Nahida Desktop/${appVersion}`,
-        };
-    }
-
-    public async fetcher(url: string, options?: RequestInit) {
-        const prefixs = ["http://localhost", "https://api.nahida.live"];
-        const isNHD = prefixs.some((prefix) => url.startsWith(prefix));
-
-        const resp = await ky(url, {
-            ...options,
-            headers: {
-                ...options?.headers,
-                ...(await this.getHeaders(url)),
-            },
-            timeout: 100000,
-            retry: {
-                limit: 2,
-            },
-            // @ts-expect-error - dispatcher is not in the type definition, but it's passed through to fetch.
-            dispatcher: await this.getAgent(),
-            hooks: {
-                afterResponse: [
-                    async (_req, _opt, resp) => {
-                        if (resp.status === 401 && isNHD) {
-                            await this.service.auth.getSession();
-                        }
-                    },
-
-                    (_request, _options, response) => {
-                        if (response.status === 524) {
-                            return new Response("cloudflare timeout. but it's ok", { status: 200 });
-                        } else {
-                            return response;
-                        }
-                    },
-                ],
-
-                beforeError: [
-                    async (error) => {
-                        const { response } = error;
-                        if (response && response.status === 524) {
-                            return error;
-                        }
-
-                        return error;
-                    },
-                ],
-            },
-        });
-
-        return resp;
+        await this.httpService.updateProxy();
     }
 }
