+++ src/main/lib/custom-downloader.ts
@@ -17,8 +17,8 @@ export class CustomDownloader {
         this.desktop = desktop;
         this.downloader = new ParallelDownloader({
             logger: desktop.logger,
-            getAgent: () => this.desktop.getAgent(),
-            getHeaders: (url: string) => this.desktop.getHeaders(url),
+            getAgent: () => this.desktop.httpService.getAgent(),
+            getHeaders: (url: string) => this.desktop.httpService.getHeaders(url),
         });
     }
 
@@ -58,9 +58,9 @@ export class CustomDownloader {
                         }
                     }
                 },
-                headers: await this.desktop.getHeaders(url),
+                headers: await this.desktop.httpService.getHeaders(url),
                 // @ts-expect-error - dispatcher is not in the type definition, but it's passed through to fetch.
-                dispatcher: await this.desktop.getAgent(),
+                dispatcher: await this.desktop.httpService.getAgent(),
             });
             if (!resp.ok) {
                 throw new Error(`Failed to download file: ${resp.statusText}`);
@@ -239,9 +239,9 @@ export class CustomDownloader {
         const resp = await ky.head(fileUrl, {
             redirect: "follow",
             throwHttpErrors: false,
-            headers: await this.desktop.getHeaders(fileUrl),
+            headers: await this.desktop.httpService.getHeaders(fileUrl),
             // @ts-expect-error - dispatcher is not in the type definition, but it's passed through to fetch.
-            dispatcher: await this.desktop.getAgent(),
+            dispatcher: await this.desktop.httpService.getAgent(),
         });
         if (!resp.ok) {
             throw new Error(`Failed to get real file URL: ${resp.statusText}`);
